Removed ProxyClient and combined into Client.
